+++ bloop/__init__.py
@@ -34,7 +34,6 @@ from .types import (
     String,
     Timestamp,
 )
-
 from .util import missing