+++ tests/integ/conftest.py
@@ -1,25 +1,18 @@
-import contextlib
-import hashlib
-import os
 import random
-import shutil
-import socket
 import string
 import subprocess
-import zipfile
 
 import boto3
 import pytest
-import requests
 from tests.helpers.utils import get_tables
 
 from bloop import BaseModel, BloopException, Engine
 from bloop.session import SessionWrapper
 from bloop.util import walk_subclasses
 
-
-LATEST_DYNAMODB_LOCAL_SHA = "70d9a92529782ac93713258fe69feb4ff6e007ae2c3319c7ffae7da38b698a61"
-DYNAMODB_LOCAL_SINGLETON = None
+DOCKER_START_COMMAND = ["docker", "run", "-d", "-p", "8000:8000", "--name", "ddb-local", "amazon/dynamodb-local"]
+DOCKER_STOP_COMMAND = ["docker", "stop", "ddb-local"]
+DOCKER_RM_COMMAND = ["docker", "rm", "ddb-local"]
 
 
 class PatchedDynamoDBClient:
@@ -29,19 +22,16 @@ class PatchedDynamoDBClient:
     def describe_time_to_live(self, **_):
         return {"TimeToLiveDescription": {"TimeToLiveStatus": "DISABLED"}}
 
+    def describe_continuous_backups(self, **_):
+        return {"ContinuousBackupsDescription": {"ContinuousBackupsStatus": "DISABLED"}}
+
     def __getattr__(self, name):
         return getattr(self.__client, name)
 
 
 class DynamoDBLocal:
-    def __init__(self, localdir: str) -> None:
-        self.localdir = localdir
-        self.process = None  # type: subprocess.Popen
-        self.port = None
-
-    @property
-    def running(self) -> bool:
-        return self.process is not None
+    def __init__(self) -> None:
+        self.running = False
 
     @property
     def session(self) -> boto3.Session:
@@ -55,7 +45,7 @@ class DynamoDBLocal:
     @property
     def endpoint(self) -> str:
         assert self.running
-        return "http://localhost:" + str(self.port)
+        return "http://localhost:8000"
 
     @property
     def clients(self) -> tuple:
@@ -71,73 +61,14 @@ class DynamoDBLocal:
 
     def start(self) -> None:
         assert not self.running
-        self._download()
-        self._reserve_port()
-        self.process = self._run()
+        self.running = True
+        subprocess.run(DOCKER_START_COMMAND, stdout=subprocess.PIPE, check=True)
 
     def stop(self) -> None:
         assert self.running
-        self.process.terminate()
-
-    def _download(self) -> None:
-        if os.path.exists(self.localdir):
-            return
-        print("\n".join((
-            "*" * 79,
-            "DynamoDBLocal doesn't exist, installing at {}".format(self.localdir),
-            "*" * 79
-        )))
-
-        # need a temp directory to download it in...
-        tempdir = self.localdir + ".tmp"
-        if os.path.exists(tempdir):
-            shutil.rmtree(tempdir)
-        os.mkdir(tempdir)
-
-        r = requests.get("https://s3-us-west-2.amazonaws.com/dynamodb-local/dynamodb_local_latest.zip",
-                         stream=True)
-        dist = os.path.join(tempdir, "dynamodb_local_latest.zip")
-
-        # download in chucks, checking its sha256 hash along the way
-        sha = hashlib.sha256()
-        with open(dist, "wb") as f:
-            for chunk in r.iter_content(chunk_size=1024):
-                if chunk:
-                    f.write(chunk)
-                    sha.update(chunk)
-
-        # Is this the file we're looking for?
-        if sha.hexdigest() != LATEST_DYNAMODB_LOCAL_SHA:
-            msg = "Invalid hash of {}/dynamodb_local_latest.zip -- expected {} but was {}"
-            raise RuntimeError(msg.format(tempdir, LATEST_DYNAMODB_LOCAL_SHA, sha.hexdigest()))
-
-        zip_ref = zipfile.ZipFile(dist, "r")
-        zip_ref.extractall(tempdir)
-        zip_ref.close()
-
-        # clean up
-        os.rename(tempdir, self.localdir)
-
-    def _reserve_port(self) -> None:
-        """
-        By using a socket and binding to "" (localhost), with a port of 0 (socket picks first non-privileged port
-        that is not in use, we can ensure that the port is available.
-        See:  https://stackoverflow.com/questions/2838244/get-open-tcp-port-in-python/2838309#2838309
-        and   https://stackoverflow.com/a/45690594
-        :return: The port to use
-        """
-        with contextlib.closing(socket.socket(socket.AF_INET, socket.SOCK_STREAM)) as s:
-            s.bind(("", 0))
-            s.listen(1)
-            self.port = s.getsockname()[1]
-
-    def _run(self) -> subprocess.Popen:
-        return subprocess.Popen([
-            "java", "-Djava.library.path=./DynamoDBLocal_lib",
-            "-jar", "DynamoDBLocal.jar", "-inMemory",
-            "-port", str(self.port)],
-            cwd=self.localdir
-        )
+        subprocess.run(DOCKER_STOP_COMMAND, stdout=subprocess.PIPE, check=True)
+        subprocess.run(DOCKER_RM_COMMAND, stdout=subprocess.PIPE, check=True)
+        self.running = False
 
 
 def pytest_addoption(parser):
@@ -147,7 +78,7 @@ def pytest_addoption(parser):
         help="make table names unique for parallel runs")
     parser.addoption(
         "--skip-cleanup", action="store_true", default=False,
-        help="don't clean up tables after tests run")
+        help="don't clean up the docker instance after tests run")
 
     default_localdir = ".dynamodb-local"
     parser.addoption(
@@ -159,31 +90,31 @@ def pytest_addoption(parser):
 @pytest.fixture(scope="session")
 def dynamodb_local(request):
     nonce = request.config.getoption("--nonce")
-    localdir = request.config.getoption("--dynamodb-local-dir")
     skip_cleanup = request.config.getoption("--skip-cleanup")
 
-    dynamodb_local = DynamoDBLocal(localdir)
+    dynamodb_local = DynamoDBLocal()
     dynamodb_local.start()
 
     yield dynamodb_local
 
+    if skip_cleanup:
+        print("Skipping cleanup, leaving docker image intact")
+        return
     try:
-        if skip_cleanup:
-            print("Skipping cleanup")
-        else:
-            print("Cleaning up tables with nonce '{}'".format(nonce))
-            dynamodb, _ = dynamodb_local.clients
-            tables = get_tables(dynamodb)
-            for table in tables:
-                if nonce not in table:
-                    continue
-                # noinspection PyBroadException
-                try:
-                    print("Removing table: {}".format(table))
-                    dynamodb.delete_table(TableName=table)
-                except Exception:
-                    print("Failed to clean up table '{}'".format(table))
+        print("Cleaning up tables with nonce '{}'".format(nonce))
+        dynamodb, _ = dynamodb_local.clients
+        tables = get_tables(dynamodb)
+        for table in tables:
+            if nonce not in table:
+                continue
+            # noinspection PyBroadException
+            try:
+                print("Removing table: {}".format(table))
+                dynamodb.delete_table(TableName=table)
+            except Exception:
+                print("Failed to clean up table '{}'".format(table))
     finally:
+        print("Shutting down ddb-local")
         dynamodb_local.stop()